Setup mongo schema and connection
