+++ docker-compose.yaml
@@ -6,3 +6,12 @@ services:
     container_name: client
     ports:
       - 8080:80
+
+  mongo:
+    image: mongo:4.2
+    container_name: mongo
+    ports:
+    - 27017:27017
+    environment:
+      MONGO_INITDB_ROOT_USERNAME: mongo
+      MONGO_INITDB_ROOT_PASSWORD: mongo
