+++ server/functions/package.json
@@ -22,6 +22,7 @@
     "firebase-functions": "^3.0.1",
     "graphql": "^14.0.2",
     "lodash": "^4.17.11",
+    "mongoose": "^5.9.2",
     "nodemailer": "^6.2.1",
     "request": "^2.88.0",
     "request-promise": "^4.2.4",
